requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
PyMySQL==1.1.0
qrcode==7.4.2
Pillow==10.0.1
numpy>=1.21.0
python-dotenv==1.0.0
cryptography==41.0.7
email-validator==2.1.0
geopy==2.3.0
requests==2.31.0
waitress==3.0.0
pandas==2.0.3
openpyxl==3.1.2
reportlab==4.0.4 